Refactor to have scopes passed through an auth provider instead of GraphSession. Moves AuthorizationHandler and TokenCredentialAuthProvider to the same module
