+++ msgraphcore/middleware/authorization.py
@@ -1,4 +1,4 @@
-from ._base_auth import AuthProviderBase
+from ._base_auth import AuthProviderBase, TokenCredential
 from ..constants import AUTH_MIDDLEWARE_OPTIONS
 from ._middleware import BaseMiddleware
 
@@ -26,3 +26,12 @@ class AuthorizationHandler(BaseMiddleware):
 
     def _get_middleware_options(self, request):
         return request.middleware_control.get(AUTH_MIDDLEWARE_OPTIONS)
+
+
+class TokenCredentialAuthProvider(AuthProviderBase):
+    def __init__(self, scopes: str, credential: TokenCredential):
+        self.credential = credential
+        self.scopes = scopes
+
+    def get_access_token(self):
+        return self.credential.get_token(self.scopes)[0]
