--- msgraphcore/middleware/authorization_provider.py
@@ -1,10 +1,0 @@
-from ._base_auth import AuthProviderBase, TokenCredential
-
-
-class TokenCredentialAuthProvider(AuthProviderBase):
-    def __init__(self, scopes: str, credential: TokenCredential):
-        self.credential = credential
-        self.scopes = scopes
-
-    def get_access_token(self):
-        return self.credential.get_token(self.scopes)[0]
